+++ full-moon/src/ast/mod.rs
@@ -495,6 +495,22 @@ pub enum Stmt {
     /// Only available when the "luau" feature flag is enabled.
     #[cfg(feature = "luau")]
     TypeFunction(TypeFunction),
+    /// A declared global, such as `declare game: DataModel`
+    /// Only available when the "luau" feature flag is enabled.
+    #[cfg(feature = "luau")]
+    DeclaredGlobal(DeclaredGlobal),
+    /// A declared global function, such as `declare function foo(a: number)`
+    /// Only available when the "luau" feature flag is enabled.
+    #[cfg(feature = "luau")]
+    DeclaredGlobalFunction(DeclaredGlobalFunction),
+    /// A declared extern type, such as `declare extern type Foo with end`
+    /// Only available when the "luau" feature flag is enabled.
+    #[cfg(feature = "luau")]
+    DeclaredExternType(DeclaredExternType),
+    /// A declared class, such as `declare class Foo end`
+    /// Only available when the "luau" feature flag is enabled.
+    #[cfg(feature = "luau")]
+    DeclaredClass(DeclaredClass),
 
     /// A goto statement, such as `goto label`
     /// Only available when the "lua52" or "luajit" feature flag is enabled.