+++ full-moon/tests/luau_declarations_cases/fail/parser/declared_extern_type_method_missing_self/source.lua
@@ -1,0 +1,3 @@
+declare extern type Foo with
+    function foo(a: number)
+end