+++ full-moon/tests/luau_declarations_cases/pass/declared_class/source.lua
@@ -1,0 +1,6 @@
+declare class Foo
+    x: number
+end
+
+declare class Bar extends Foo
+end