+++ full-moon/tests/luau_declarations_cases/pass/declared_extern_type/source.lua
@@ -1,0 +1,6 @@
+declare extern type Foo with
+    function foo(self, x: number)
+    @deprecated function bar(self)
+    ["baz"]: string
+    [string]: any
+end