+++ full-moon/tests/luau_declarations_cases/pass/declared_functions/source.lua
@@ -1,0 +1,4 @@
+declare function foo(a: string, b: boolean, ...: number): number
+
+@deprecated
+declare function bar(b: number)